requirements.txt
# 深度学习框架
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0
# sentence-transformers>=2.2.0  # 已移除，使用transformers替代

# 数据处理
numpy>=1.21.0
pandas>=1.4.0
scikit-learn>=1.1.0
Pillow>=9.0.0
opencv-python>=4.6.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# 配置和日志
PyYAML>=6.0
loguru>=0.6.0

# 文件处理
PyPDF2>=2.10.0
pdfplumber>=0.7.0

# 工具库
tqdm>=4.64.0
typing-extensions>=4.3.0

# 可选依赖（用于高级功能）
# 如果需要使用CLIP模型
# clip-by-openai>=1.0

# 如果需要使用更多预训练模型
# timm>=0.6.0

# 如果需要使用Weights & Biases进行实验跟踪
# wandb>=0.13.0

# 如果需要使用TensorBoard
# tensorboard>=2.10.0

# 如果需要使用更高级的数据增强
# albumentations>=1.2.0

# 如果需要使用分布式训练
# accelerate>=0.12.0